Revert "refactor: Avoid ambiguity problems with clippy"
This reverts commit 50b54ea3b1a1072046909cf971938cbdbbcbabca.
